Decompose the unit-if and option-match lowering units out of control_b
